add restriction to post deletion and editing only for creators

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -41,20 +42,27 @@ exports.createPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     const imageUrl = req.file.path.replace("\\" ,"/");
-    console.log(title, content, 'req.body');
+    let creator;
     const post = new Post({
         title, 
         content, 
         imageUrl: imageUrl,
-        creator: {
-            name: 'Igor',
-        },
+        creator: req.userId,
     });
     post.save()
+    .then(res => {
+       return User.findById(req.userId)
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
     .then(result => {
         res.status(201).json({
             message: 'Post created successfully',
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         });
     })
     .catch(err => {
@@ -111,6 +119,12 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            // Only creator can deleted its post check
+            if (post.creator.toString() !== req.userId.toString()) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -139,12 +153,25 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            // Only creator can deleted its post check
+            if (post.creator.toString() !== req.userId.toString()) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndDelete(postId);
         })
-        .then(result => {
-            console.log(result);
+        .then(() => {
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            // Inbuilt mongoose filter method to clear relation
+            user.posts.pull(postId)
+            user.save();
+        })
+        .then(() => {
             res.status(200).json({ message: 'Deleted Post.' })
         })
         .catch(err => {
